requirements.txt
asgiref==3.5.2
cached-property==1.5.2
Django==4.0.5
django-cors-headers==3.13.0
django-daterangefilter==1.0.0
django-filter==21.1
django-url-filter==0.3.15
djangorestframework==3.13.1
drf-url-filters==0.5.1
enum-compat==0.0.3
importlib-metadata==4.11.4
Markdown==3.3.7
Pillow==9.1.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
sweetify==2.3.1
tzdata==2022.1
voluptuous==0.13.1
zipp==3.8.0
